renamed post action, login action and moved login and logout controllers into authentication controller
