--- module/Application/src/Controller/LogoutController.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace Application\Controller;
-
-use Zend\Mvc\Controller\AbstractActionController;
-use Zend\View\Model\ViewModel;
-
-
-class LogoutController extends AbstractActionController
-{
-    const LOGIN_HEADER = 'Location: /login';
-
-    public function logoutAction() {
-        session_destroy();
-        header(self::LOGIN_HEADER);
-        exit;
-    }
-}
