+++ module/Application/src/Controller/UserAuthenticationController.php
@@ -12,17 +12,18 @@ namespace Application\Controller;
 use Zend\Mvc\Controller\AbstractActionController;
 use Zend\View\Model\ViewModel;
 
-class LoginController extends AbstractActionController
+class UserAuthenticationController extends AbstractActionController
 {
     private $user;
     const ACCOUNT_HEADER = 'Location: /account';
+    const LOGIN_HEADER = 'Location: /login';
 
     public function __construct($user)
     {
         $this->user = $user;
     }
 
-    public function loginAction() {
+    public function indexAction() {
         if(!empty($_SESSION['userAuth'])) {
             if($this->user->validateToken($_SESSION['userAuth'], $_SESSION['id']))
             {
@@ -36,5 +37,11 @@ class LoginController extends AbstractActionController
         } else {
             return new ViewModel();
         }
+    }
+
+    public function logoutAction() {
+        session_destroy();
+        header(self::LOGIN_HEADER);
+        exit;
     }
 }
