+++ module/Application/view/application/account/index.phtml
@@ -1,4 +1,4 @@
 <div class="container" id="test">
     <h1>Account</h1>
-    <a href="logout" id="logout" class="btn btn-danger">Logout</a>
+    <a href="/logout" id="logout" class="btn btn-danger">Logout</a>
 </div>
